+++ packages/dropdown_button2/lib/src/dropdown_menu_item.dart
@@ -100,6 +100,9 @@ class _DropdownMenuItemContainer extends StatelessWidget {
   /// If set to true, then this item's height will vary according to its
   /// intrinsic height instead of using [height] property.
   ///
+  /// It is highly recommended to keep this value as false when dealing with
+  /// a significantly large items list in order to optimize performance.
+  ///
   /// Note: If set to true and there isn't enough vertical room for the menu, there's
   /// no way to know the item's intrinsic height in-advance to properly scroll to
   /// the selected item. Instead, the provided [height] value will be used, which means
@@ -147,7 +150,6 @@ class _DropdownItemButtonState<T> extends State<_DropdownItemButton<T>> {
     switch (FocusManager.instance.highlightMode) {
       case FocusHighlightMode.touch:
         inTraditionalMode = false;
-        // TODO(Ahmed): Remove decorative breaks and add lint to it [flutter>=v3.10.0].
         break;
       case FocusHighlightMode.traditional:
         inTraditionalMode = true;
@@ -173,6 +175,7 @@ class _DropdownItemButtonState<T> extends State<_DropdownItemButton<T>> {
     final DropdownItem<T> dropdownItem = widget.route.items[widget.itemIndex];
 
     dropdownItem.onTap?.call();
+    widget.route.onChanged?.call(dropdownItem.value);
 
     if (dropdownItem.closeOnTap) {
       Navigator.pop(
@@ -201,8 +204,8 @@ class _DropdownItemButtonState<T> extends State<_DropdownItemButton<T>> {
     final DropdownItem<T> dropdownItem = widget.route.items[widget.itemIndex];
     final double unit = 0.5 / (widget.route.items.length + 1.5);
     final double start =
-        _clampDouble(menuCurveEnd + (widget.itemIndex + 1) * unit, 0.0, 1.0);
-    final double end = _clampDouble(start + 1.5 * unit, 0.0, 1.0);
+        clampDouble(menuCurveEnd + (widget.itemIndex + 1) * unit, 0.0, 1.0);
+    final double end = clampDouble(start + 1.5 * unit, 0.0, 1.0);
     final CurvedAnimation opacity = CurvedAnimation(
         parent: widget.route.animation!, curve: Interval(start, end));